+++ src/main/java/com/example/praktikalistproducts/dao/impl/PostgresProductDAO.java
@@ -174,11 +174,11 @@ public class PostgresProductDAO implements ProductDAO {
      * @throws RuntimeException если произошла ошибка при удалении
      */
     @Override
-    public void deleteProduct(int id) {
+    public void deleteProduct(long id) {
         String sql = "DELETE FROM products WHERE id = ?";
 
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
-            pstmt.setInt(1, id);
+            pstmt.setInt(1, (int) id);
             pstmt.executeUpdate();
         } catch (SQLException e) {
             throw new RuntimeException("Ошибка при удалении продукта", e);
